+++ composeApp/src/commonMain/kotlin/com/majotyler/hiittimer/presentation/playWorkoutScreen/PlayWorkoutScreen.kt
@@ -36,6 +36,7 @@ fun PlayWorkoutScreen(
     val play by viewModel.play.collectAsStateWithLifecycle()
     val workoutPlayState by viewModel.workoutPlayState.collectAsStateWithLifecycle()
     val enabled by viewModel.enabled.collectAsStateWithLifecycle()
+    val workoutCompleted by viewModel.workoutCompleted.collectAsStateWithLifecycle()
 
     BackHandler(
         onBack = {
@@ -62,6 +63,8 @@ fun PlayWorkoutScreen(
             )
         },
         enabled = enabled,
+        workoutCompleted = workoutCompleted,
+        onClickedSeeWorkout = { viewModel.onEvent(event = PlayWorkoutViewEvent.ClickedSeeWorkout) },
     )
 }
 
@@ -77,6 +80,8 @@ private fun PlayWorkoutContent(
     onClickedDialogCancel: () -> Unit,
     onClickedPause: () -> Unit,
     enabled: Boolean,
+    workoutCompleted: Boolean,
+    onClickedSeeWorkout: () -> Unit,
 ) {
     if (confirmationDialogVisible) {
         ConfirmationDialog(
@@ -130,8 +135,12 @@ private fun PlayWorkoutContent(
         }, enabled = enabled) {
             Text(text = buttonLabel)
         }
-
-
+        if (workoutCompleted) {
+            Spacer(modifier = Modifier.height(12.dp))
+            Button(onClick = onClickedSeeWorkout) {
+                Text("See Workout")
+            }
+        }
     }
 
 }
