+++ composeApp/src/commonMain/kotlin/com/majotyler/hiittimer/presentation/workoutReviewScreen/WorkoutReviewDestination.kt
@@ -1,0 +1,7 @@
+package com.majotyler.hiittimer.presentation.workoutReviewScreen
+
+import com.majotyler.hiittimer.presentation.common.navigation.Destination
+
+sealed interface WorkoutReviewDestination : Destination {
+    data object NavigateUp : WorkoutReviewDestination
+}
