+++ composeApp/src/commonMain/kotlin/com/majotyler/hiittimer/presentation/workoutReviewScreen/WorkoutReviewViewEvent.kt
@@ -1,0 +1,7 @@
+package com.majotyler.hiittimer.presentation.workoutReviewScreen
+
+sealed interface WorkoutReviewViewEvent {
+    data object ClickedConnectWithStrava : WorkoutReviewViewEvent
+    data object ClickedCreateStravaActivity : WorkoutReviewViewEvent
+    data object ClickedNavigateUp : WorkoutReviewViewEvent
+}
